device/container/requirements.txt
requests==2.24.0
paho-mqtt==1.5.0
pyopenssl==19.1.0